# Internalize types and harmonize parameters and access