--- a/Fluid.Tests/Directory.Build.props
+++ b/Fluid.Tests/Directory.Build.props
@@ -1,0 +1,5 @@
+﻿<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+
+  <Import Project="../Common.props"/>
+
+</Project>