--- a/Fluid/Accessors/AsyncDelegateAccessor.cs
+++ b/Fluid/Accessors/AsyncDelegateAccessor.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 
 namespace Fluid.Accessors
 {
-    public class AsyncDelegateAccessor : AsyncDelegateAccessor<object, object>
+    internal sealed class AsyncDelegateAccessor : AsyncDelegateAccessor<object, object>
     {
         public AsyncDelegateAccessor(Func<object, string, Task<object>> getter) : base((obj, name, ctx) => getter(obj, name))
         {