--- a/Fluid/Accessors/MethodInfoAccessor.cs
+++ b/Fluid/Accessors/MethodInfoAccessor.cs
@@ -2,7 +2,7 @@
 
 namespace Fluid.Accessors
 {
-    public class MethodInfoAccessor : IMemberAccessor
+    internal sealed class MethodInfoAccessor : IMemberAccessor
     {
         private readonly MethodInfo _methodInfo;