--- a/Fluid/Ast/AssignStatement.cs
+++ b/Fluid/Ast/AssignStatement.cs
@@ -5,7 +5,7 @@ using Fluid.Values;
 
 namespace Fluid.Ast
 {
-    public class AssignStatement : Statement
+    internal sealed class AssignStatement : Statement
     {
         public AssignStatement(string identifier, Expression value)
         {