--- a/Fluid/Ast/BinaryExpressions/EqualBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/EqualBinaryExpression.cs
@@ -2,13 +2,13 @@
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public sealed class EqualBinaryExpression : BinaryExpression
+    internal sealed class EqualBinaryExpression : BinaryExpression
     {
         public EqualBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
+        protected override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
             return leftValue.Equals(rightValue)
                 ? BooleanValue.True