--- a/Fluid/Ast/BinaryExpressions/LowerThanExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/LowerThanExpression.cs
@@ -2,7 +2,7 @@
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public sealed class LowerThanExpression : BinaryExpression
+    internal sealed class LowerThanExpression : BinaryExpression
     {
         public LowerThanExpression(Expression left, Expression right, bool strict) : base(left, right)
         {
@@ -11,7 +11,7 @@ namespace Fluid.Ast.BinaryExpressions
 
         public bool Strict { get; }
 
-        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
+        protected override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
             if (leftValue.IsNil() || rightValue.IsNil())
             {