--- a/Fluid/Ast/BinaryExpressions/ModuloBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/ModuloBinaryExpression.cs
@@ -2,13 +2,13 @@
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public sealed class ModuloBinaryExpression : BinaryExpression
+    internal sealed class ModuloBinaryExpression : BinaryExpression
     {
         public ModuloBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
+        protected override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
             return leftValue is NumberValue && rightValue is NumberValue
                 ? NumberValue.Create(leftValue.ToNumberValue() % rightValue.ToNumberValue())