--- a/Fluid/Ast/BinaryExpressions/MultiplyBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/MultiplyBinaryExpression.cs
@@ -2,13 +2,13 @@
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public sealed class MultiplyBinaryExpression : BinaryExpression
+    internal sealed class MultiplyBinaryExpression : BinaryExpression
     {
         public MultiplyBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
+        protected override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
             return leftValue is NumberValue && rightValue is NumberValue
                 ? NumberValue.Create(leftValue.ToNumberValue() * rightValue.ToNumberValue())