--- a/Fluid/Ast/BinaryExpressions/OrBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/OrBinaryExpression.cs
@@ -2,13 +2,13 @@
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public sealed class OrBinaryExpression : BinaryExpression
+    internal sealed class OrBinaryExpression : BinaryExpression
     {
         public OrBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
+        protected override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
             return BooleanValue.Create(leftValue.ToBooleanValue() || rightValue.ToBooleanValue());
         }