--- a/Fluid/Ast/BreakStatement.cs
+++ b/Fluid/Ast/BreakStatement.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 namespace Fluid.Ast
 {
-    public class BreakStatement : Statement
+    internal sealed class BreakStatement : Statement
     {
         public override ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
         {