--- a/Fluid/Ast/CycleStatement.cs
+++ b/Fluid/Ast/CycleStatement.cs
@@ -1,36 +1,31 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using Fluid.Values;
 
 namespace Fluid.Ast
 {
-    public sealed class CycleStatement : Statement
+    internal sealed class CycleStatement : Statement
     {
-        private readonly Expression[] _values;
+        private readonly List<Expression> _values;
+        private readonly Expression _group;
 
-        public CycleStatement(Expression group, Expression[] values)
+        public CycleStatement(Expression group, params Expression[] values) : this(group, new List<Expression>(values))
         {
-            Group = group;
+        }
+
+        public CycleStatement(Expression group, List<Expression> values)
+        {
+            _group = group;
             _values = values;
         }
-
-        public CycleStatement(Expression group, IList<Expression> values)
-        {
-            Group = group;
-            _values = values.ToArray();
-        }
-
-        public Expression Group { get; }
-        public IList<Expression> Values2 { get; }
 
         public override async ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
         {
             context.IncrementSteps();
 
-            var groupValue = Group == null ? "$cycle_" : "$cycle_" + (await Group.EvaluateAsync(context)).ToStringValue();
+            var groupValue = _group == null ? "$cycle_" : "$cycle_" + (await _group.EvaluateAsync(context)).ToStringValue();
 
             var currentValue = context.GetValue(groupValue);
 
@@ -39,7 +34,7 @@ namespace Fluid.Ast
                 currentValue = NumberValue.Zero;
             }
 
-            var index = (uint) currentValue.ToNumberValue() % _values.Length;
+            var index = (int) currentValue.ToNumberValue() % _values.Count;
             var value = await _values[index].EvaluateAsync(context);
             context.SetValue(groupValue, NumberValue.Create(index + 1));