--- a/Fluid/Ast/DecrementStatement.cs
+++ b/Fluid/Ast/DecrementStatement.cs
@@ -5,14 +5,14 @@ using Fluid.Values;
 
 namespace Fluid.Ast
 {
-    public class DecrementStatement : Statement
+    internal sealed class DecrementStatement : Statement
     {
+        private readonly string _identifier;
+
         public DecrementStatement(string identifier)
         {
-            Identifier = identifier;
+            _identifier = identifier;
         }
-
-        public string Identifier { get; }
 
         public override ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
         {
@@ -22,11 +22,11 @@ namespace Fluid.Ast
             // Variable identifiers don't represent the same slots as inc/dec ones.
             // c.f. https://shopify.github.io/liquid/tags/variable/
 
-            var prefixedIdentifier = IncrementStatement.Prefix + Identifier;
+            var prefixedIdentifier = IncrementStatement.Prefix + _identifier;
 
             var value = context.GetValue(prefixedIdentifier);
-            
-            if (value.IsNil()) 
+
+            if (value.IsNil())
             {
                 value = NumberValue.Zero;
             }