--- a/Fluid/Ast/IdentifierSegment.cs
+++ b/Fluid/Ast/IdentifierSegment.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 namespace Fluid.Ast
 {
     [DebuggerDisplay("{Identifier,nq}")]
-    public class IdentifierSegment : MemberSegment
+    internal sealed class IdentifierSegment : MemberSegment
     {
         public IdentifierSegment(string identifier)
         {