--- a/Fluid/Ast/IndexerSegment.cs
+++ b/Fluid/Ast/IndexerSegment.cs
@@ -3,18 +3,18 @@ using Fluid.Values;
 
 namespace Fluid.Ast
 {
-    public class IndexerSegment : MemberSegment
+    internal sealed class IndexerSegment : MemberSegment
     {
+        private readonly Expression _expression;
+
         public IndexerSegment(Expression expression)
         {
-            Expression = expression;
+            _expression = expression;
         }
-
-        public Expression Expression { get; }
 
         public override async ValueTask<FluidValue> ResolveAsync(FluidValue value, TemplateContext context)
         {
-            var index = await Expression.EvaluateAsync(context);
+            var index = await _expression.EvaluateAsync(context);
             return await value.GetIndexAsync(index, context);
         }
     }