* internalize all types that can be internalized
* seal all types that can be sealed
* consistently take List<T> as parameter to expressions and statements
* add `AsSpan` helper for faster List<T> iteration, accesses the internal array directly (but cannot be used in async context, because Span is ref struct, works in many places where we try to avoid awaits)
* used InternalsVisible to allow some type/member access from test and view engine projects

Tested compilation with OrchardCore, NJsonSchema and FluentEmail. All these projects still compile with these changes after hiding and sealing members.

Please change this PR directly if you find types that you think should be public/non-sealed.

## Benchmarks

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  DefaultJob : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT


```

### Before

|         Method |       Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |-----------:|----------:|----------:|-------:|-------:|----------:|
|          Parse |   4.519 μs | 0.0131 μs | 0.0123 μs | 0.1526 |      - |      3 KB |
|       ParseBig |  23.141 μs | 0.0473 μs | 0.0395 μs | 0.7019 | 0.0305 |     12 KB |
|         Render | 222.656 μs | 0.2196 μs | 0.2054 μs | 5.6152 | 1.2207 |     96 KB |
| ParseAndRender | 225.082 μs | 0.8097 μs | 0.7574 μs | 5.8594 | 1.2207 |     99 KB |

### After

|         Method |       Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |-----------:|----------:|----------:|-------:|-------:|----------:|
|          Parse |   4.472 μs | 0.0059 μs | 0.0052 μs | 0.1526 |      - |      3 KB |
|       ParseBig |  23.116 μs | 0.0611 μs | 0.0510 μs | 0.7019 | 0.0305 |     11 KB |
|         Render | 221.587 μs | 0.4490 μs | 0.4200 μs | 5.6152 | 1.2207 |     96 KB |
| ParseAndRender | 219.165 μs | 0.4073 μs | 0.3810 μs | 5.8594 | 1.2207 |     99 KB |